+++ closure/goog/async/nexttick_test.html
@@ -9,12 +9,12 @@ See the COPYING file for details.
  <head>
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <title>
-   Closure Unit Tests - goog.async.nextTick
+   Closure Unit Tests - goog.async.NextTick
   </title>
   <script src="../base.js">
   </script>
   <script>
-   goog.require('goog.async.nextTickTest');
+   goog.require('goog.async.NextTickTest');
   </script>
  </head>
  <body>
